refactor: move sort_batches_and_print to test_util
