+++ src/mito2/src/test_util.rs
@@ -14,6 +14,7 @@
 
 //! Utilities for testing.
 
+pub mod batch_util;
 pub mod memtable_util;
 pub mod meta_util;
 pub mod scheduler_util;
